# openapi.yaml
openapi: 3.0.3
info:
  version: '1.0'
  title: Reactive Bank Account
servers:
  - url: http://localhost:8080
tags:
  - name: Пользователь
    description: Отвечает за работу с пользователями
  - name: Счет
    description: Отвечает за работу с банковским счетом
paths:
  /api/v1/user:
    get:
      tags:
        - Пользователь
      summary: Получение профиля пользователя
      description: Позволяет получить профиль пользователя
      operationId: getUserProfile
      responses:
        '200':
          description: Пользователь успешно получил информацию о профиле
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Что-то пошло не так
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
      security:
        - bearerAuth:
            - type:http
            - scheme:bearer
  /api/v1/user/register:
    post:
      tags:
        - Пользователь
      summary: Регистрация пользователя
      description: Позволяет зарегистрироваться
      operationId: registerUser
      requestBody:
        description: Создание пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequestDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/RegisterUserRequestDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterUserRequestDto'
        required: true
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'          
            application/xml:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Пользователь с указанным номером телефона уже существует
        '500':
          description: Что-то пошло не так
  /api/v1/user/login:  
    post:
      tags:
        - Пользователь
      summary: Авторизация
      description: Позволяет пользователю войти в аккаунт
      operationId: loginUser
      requestBody:
        description: Авторизация пользователя
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequestDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/LoginUserRequestDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginUserRequestDto'
        required: true
      responses:
        '200':
          description: Пользователь успешно вошел в аккаунт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          description: Неверный логин или пароль
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Что-то пошло не так
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
  /api/v1/account:
    get:
      tags:
        - Счет
      summary: Получение счетов
      description: Позволяет пользователю получить его счета
      operationId: getAccount
      responses:
        '200':
          description: Счета успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccountDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BankAccountDto'
        '401':
          description: Неавторизован
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'  
      security:
        - bearerAuth:
          - type:http
          - scheme:bearer
    post:
      tags:
        - Счет
      summary: Создание счета
      description: Позволяет пользователю открыть новый счет
      operationId: createAccount
      responses:
        '200':
          description: Счет успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/BankAccountDto'
      security:
        - bearerAuth:
            - type:http
            - scheme:bearer
  /api/v1/account/transfer:
    post:
      tags:
        - Счет
      summary: Перевод денег
      description: Позволяет пользователю перевести деньги на другой счет
      operationId: sendMoney
      parameters:
        - name: accountIdFrom
          in: query
          description: Id счета, откуда перевести деньги
          required: true
          schema:
            type: string
            format: uuid
        - name: accountIdWhere
          in: query
          description: Id счета, куда перевести деньги
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Деньги успешно переведены
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/BankAccountDto'          
            application/xml:
              schema:
                $ref: '#/components/schemas/BankAccountDto'   
        '400':
          description: Неверный id счета
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'  
        '401':
          description: Не авторизован
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/Response'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'  
      security:
        - bearerAuth:
          - type:http
          - scheme:bearer
  /api/v1/account/history:
    get:
      tags:
        - Счет
      summary: Получение истории операций счета
      description: Позволяет пользователю получить историю операций своего счета
      operationId: accountOperationsHistory
      parameters:
        - name: accountId
          in: query
          description: Id счета
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: История операций успешно получена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortTransactionDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortTransactionDto'
        '404':
          description: Счет с указанным идентификатором не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
      security:
        - bearerAuth:
            - type:http
            - scheme:bearer
  /api/v1/account/transaction/{transactionId}:
    get:
      tags:
        - Счет
      summary: Получение информации о транзакции
      description: Позволяет пользователю получить подробную информацию о проведенной транзакции
      operationId: transactionInfo
      parameters:
        - name: transactionId
          in: path
          description: Id транзакции
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о транзакции успешно получена
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDto'
        '404':
          description: Транзакция с указанным идентификатором не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
            application/xml:
              schema:
                $ref: '#/components/schemas/Response'
      security:
        - bearerAuth:
            - type:http
            - scheme:bearer
components:
  schemas:
    LoginUserRequestDto:
      type: object
      properties:
        login:
          description: Логин пользователя
          type: string
          example: example
        password:
          description: Пароль пользователя
          type: string
          example: password1
    RegisterUserRequestDto:
      type: object
      properties:
        login:
          description: Логин пользователя
          type: string
          example: example
        email:
          description: Email пользователя
          type: string
          format: email
        password:
          description: Пароль пользователя
          type: string
          example: password1
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a84876e7-2795-4531-9480-4a23542f7c31
        email:
          type: string
          example: user@email.com
        login:
          type: string
          example: 'user'
      xml:
        name: user
    Response:
      type: object
      properties:
        status:
          type: integer
          format: int64
          example: 200
        message:
          type: string
          example: "Все отлично"
        timestamp:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/CustomFieldError'
    CustomFieldError:
      type: object
      properties:
        field:
          type: string
          example: "login"
        message:
          type: string
          example: "Логин должен быть указан"
    BankAccountDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amountOfMoney:
          type: number
          format: double
    TransactionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountIdFrom:
          type: string
          format: uuid
        accountIdWhere:
          type: string
          format: uuid
        amount:
          type: string
          format: decimal
        timestamp:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/TransactionEnum'
    ShortTransactionDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: string
          format: decimal
        timestamp:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/TransactionEnum'
    TransactionEnum:
      type: string
      enum:
        - REFILL
        - TRANSFER
        - DEPOSIT
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer