# src/main/kotlin/com/learning/reactive/dto/transaction/TransactionEnum.kt
package com.learning.reactive.dto.transaction

/**
 * Enum для отображения статуса транзакции
 */
enum class TransactionEnum {
    /**
     * Пополнение
      */
    REFILL,

    /**
     * Перевод
     */
    TRANSFER,

    /**
     * Пополнение денег из банкомата
     */
    DEPOSIT
}